% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tucker3Package.R
\name{Imputemethod}
\alias{Imputemethod}
\title{Imputemethod  –  Tucker-3 iterative imputation (Imputation of missing values via iterative Tucker-3)}
\usage{
Imputemethod(
  X,
  fac = c(2, 2, 2),
  conver = 1e-07,
  max.iter = 1000,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{X}{3-way array (subjects × variables × time)}

\item{fac}{integer length-3: n° components for each mode (default c(2,2,2))}

\item{conver}{defauld threshold convergence ΔSSTc (default 1e-7)}

\item{max.iter}{max number of iteration (default 1000)}

\item{seed}{integer}

\item{verbose}{default(TRUE)}
}
\value{
The array \code{X} with \code{NA} replaced
}
\description{
Imputemethod  –  Tucker-3 iterative imputation (Imputation of missing values via iterative Tucker-3)
}
