% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nplsdaVips.R
\name{nplsda_vips}
\alias{nplsda_vips}
\title{N-way Partial Least Squares Discriminant Analysis -}
\usage{
nplsda_vips(X, Y, ncomp = 3, outcome.Y = NULL, slice_vip = FALSE)
}
\arguments{
\item{X}{Array 3D (n×p×k): eg. subjects, p features, k times}

\item{Y}{Matrice (n×q) or n x p x q where dim(q) == 1 classic NPLS-DA or  Array 3D (n × p x q) for class2 regression with q > 1}

\item{ncomp}{Number of components (2-10)}

\item{outcome.Y}{Outcome vector for  regression class 2 (mandatory if Y is from regression Class2)}

\item{slice_vip}{Boolean,  about computing nplsda_vips separately for each slice}
}
\value{
List with nplsda_vips, Q², explained variance and predictions for each component
}
\description{
Compute performance metrics for NPLS-DA. Supporta 2-10 components, multivariate Y,
and robust error handling
}
