% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMode3.R
\name{plot_nplsda_blockX_mode3}
\alias{plot_nplsda_blockX_mode3}
\title{Plot Block X Mode 3}
\usage{
plot_nplsda_blockX_mode3(factors, pcs = c(1, 2), edge = c(xlimSize, ylimSize))
}
\arguments{
\item{factors}{Object returned from compute_npls_factors
($FactorsX$Mode3)}

\item{pcs}{vector of integer containing the components to plot (default c(1,2))}

\item{edge}{vecctors containing size of border (default(c(xlimSize,ylimSize)))}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    Invisibile, dataframe of points plotted 
}\if{html}{\out{</div>}}
}
\description{
Plot Block X Mode 3
}
