% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScoresMode1.R
\name{plot_nplsda_scores}
\alias{plot_nplsda_scores}
\title{Plot NPLS-DA Scores}
\usage{
plot_nplsda_scores(
  scores_matrix,
  expl_var_matrix,
  class_vec,
  pc1 = 1,
  pc2 = 2,
  variance = FALSE,
  main_title = "NPLS-DA Score Plot"
)
}
\arguments{
\item{scores_matrix}{Matrix of scores (from NPLSDAvariates$X)}

\item{expl_var_matrix}{Matrix with variance explained (with R2.Y column)}

\item{class_vec}{Factor vector with class labels}

\item{pc1}{First component to plot (default = 1)}

\item{pc2}{Second component to plot (default = 2)}

\item{main_title}{Plot title (default = "NPLS-DA Score Plot")}
}
\description{
Plot NPLS-DA Scores
}
