% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsReg.R
\name{pls_reg}
\alias{pls_reg}
\title{PLS REGRESSION with incremental Q^2
PLS Regression with Incremental cross validation}
\usage{
pls_reg(X, Y, ncomp = 2, cv = TRUE)
}
\arguments{
\item{X}{Matrice predictors (n x p)}

\item{Y}{Matrice response (n x q)}

\item{ncomp}{Numero di componenti PLS (default: 2)}

\item{cv}{Logico, se TRUE esegue cross-validation (default: TRUE)}
}
\value{
list with components PLS, metrics of validations and predictions
}
\description{
PLS REGRESSION with incremental Q^2
PLS Regression with Incremental cross validation
}
\details{
Q² Computing:
Q²\link{h} = 1 - (PRESS\link{h} / RSS\link{h})
RSS\link{h} are residuals after h-1 components

Q²CUM : Q²cum\link{h} = 1 - ∏(1 - Q²\link{j}) per j=1:h
}
