% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2DPer.r
\name{plot_vip2d_with_groups_nogaps}
\alias{plot_vip2d_with_groups_nogaps}
\title{Plot Top-N VIP2D with Case/Control strip and permutation p-values}
\usage{
plot_vip2d_with_groups_nogaps(
  obj,
  X,
  groups,
  comp = 1,
  top_n = 15,
  threshold = 1,
  sep = "_",
  mode = c("winner", "effect"),
  case_level = NULL,
  case_col = "#D55E00",
  control_col = "#009E73",
  loser_alpha = 0.25,
  Y = NULL,
  ncomp = NULL,
  permute_R = 0,
  alpha_perm = 0.05,
  fit_fun = NULL,
  seed = NULL,
  workers = NULL,
  cap_blas_threads = FALSE,
  parallel = c("auto", "off"),
  future_packages = NULL,
  ...
)
}
\description{
Plot Top-N VIP2D with Case/Control strip and permutation p-values
}
